==> 1_OrderLines_Cleanup.ipynb <==
# orderlines.csv (ol_)
Every row represents each one of the different products involved in an order.
- id – a unique identifier for each row in this file
- id_order – corresponds to orders.order_id
- product_id – an old identifier for each product, nowadays not in use
- product_quantity – how many units of that product were purchased on that order
- sku – stock keeping unit: a unique identifier for each product
- unit_price – the unitary price (in euros) of each product at the moment of placing that order
- date – timestamp for the processing of that product

    ==============================================================================================

## Importing the data
- ``` glob-glob("file_pat") ``` --> read multi files 
- ``` pd.concat(dfs_list, ignore_index=True)```  --> create 1 df from multi dfs
- ``` pd.read_csv(path)```  --> create 1 df from a csv file

import pandas as pd
import numpy as np

pd.set_option("display.min_rows", 0) 
pd.set_option("display.max_rows", 30) 
pd.__version__

url = "https://drive.google.com/file/d/1FYhN_2AzTBFuWcfHaRuKcuCE6CWXsWtG/view?usp=sharing" 
path = "https://drive.google.com/uc?export=download&id="+url.split("/")[-2]
ol = pd.read_csv(path)

      ===============================================

## Rename Columns , Set Index
- Rules: 
    - add 2-3 chars prefix to all columns
    - change column name to CamelCase
    - shorten it as max as possible
    - the Unique column shall have ALL_CAP name
 - ``` df.columns```   , ``` df.index``` 
 - ``` df=df.rename(columns={"A": "a", "B": "c"})``` 
 - ``` df.columns = ["a","b":"x"]``` 
     - take care, renaming the columns like that will convert the NAN to some value!!
 - ``` df=df.set_index("col")```  , ``` df=df.reset_index()``` 

ol.columns

ol=ol.rename(columns={"id": "OL_ID"
                        , "id_order": "ol_ORD_ID"
                        , "product_id": "ol_ProdId"
                        , "product_quantity": "ol_ProdQnty"
                        , "sku": "ol_Sku"
                        , "unit_price": "ol_ProdUntPr"
                        , "date": "ol_ProcessDate"})

ol.columns

ol.index #hint: no need to change index

    ==============================================================================================

## Explore the data
- ``` df.shape``` , ``` df.size``` , ``` df.ndim``` 
- ``` df.sample(5)``` , ``` df.info()``` 
- Numerical : ``` df.describe()``` , ``` df.col.nlargest()``` , ``` df.col.nsmallest()``` 
- Category : ``` df.nunique()``` , ``` df.unique() ``` 

ol.shape

ol.sample(10)
## unit_price has numbers like 3.121.59 TODO: fix it to proper float

ol.info()  
## hint: there are no nulls 
## hint: date  is of type object, has to be datetime - TODO: fix required
## hint: unit_price is of type object, has to be float - TODO: fix required

ol.describe()
# hint: product_id seems like empty and shall be removed - TODO: Fix required
# hint: product_quantity 3rd quartile = min = 1, max=999! 
    # probably we shall handle orders with different quantities differently - TODO:fix required

ol.nunique() # hint: id is unique per row 
# hint: product_id is useless, shall be removed - TODO: fix required

ol.ol_ProdId.unique()
#hint: useless data


ol.ol_ProdUntPr.sort_values(ascending=False).tail(40) 
# There is negative value in ol_ProdUnitPrice,  TODO: Fix Required

ol.loc[(ol.ol_ProdUntPr == "0.00")].ol_ProdQnty.sum()
#That amount of product with price 0.00! TODO:check more about it 946/293983 = 0.3%

    ==============================================================================================

## Initial Clean
- Remember to create a copy of the df using ``` df.copy()``` 

ol_original = ol.copy()

      ===============================================

### Strip whitespaces
- ``` df.applymap(lambda x: x.strip() if isinstance(x, str) else x)```

ol = ol.applymap(lambda x: x.strip() if isinstance(x, str) else x)

      ===============================================
### Remove Duplicates Rows
- ``` df.duplicated().sum()``` 
- ``` df.loc[df.duplicated()==True]``` 
- ``` df=df.drop:duplicates() ``` 
- ``` df=df.drop:duplicates(subset=["col"])```  --> remove rows based on duplicated in specific column

ol.duplicated().sum() #hint : No duplicates

      =======================

#### Remove duplicated rows related to Unique columns
- find possible duplicates ``` len(df.col.unique()) == df.shape[0] ``` 
- get the excat value for duplicate columns ``` df.loc[df.duplicated(subset="col")]``` 
- find all columns with same value ``` df[df.col=="val"]``` 

check if ol_ID is unique

len(ol.OL_ID.unique()) == ol.shape[0]

      ===============================================

### Drop duplicate/un-necessary Columns
- ``` df=df.drop(columns=["col1","col2"])``` 

ol.describe() ## hint: all columns contain needed data. nothing to drop

drop ol_ProductId

ol = ol.drop(columns=["ol_ProdId"])

    ==============================================================================================

## Compare to original DataFrame
``` df.compare(df2)``` 

#ol.compare(ol_original)

    ==============================================================================================

## Export the cleaned DataFrame

``` df.to_pickle("file_name.pkl")``` 

ol.to_pickle("clean_Tables\OrderLines_c.pkl")
