==> 2_Products_Cleanup.ipynb <==
# Products.csv (pro_)
- sku – stock keeping unit: a unique identifier for each product
- name – product name
- desc – product description
- in_stock – whether or not the product was in stock at the moment of the data extraction
- type – a numerical code for product type

    ==============================================================================================

## Importing the data
- ``` glob-glob("file_pat") ``` --> read multi files 
- ``` pd.concat(dfs_list, ignore_index=True)```  --> create 1 df from multi dfs
- ``` pd.read_csv(path)```  --> create 1 df from a csv file

import pandas as pd
import numpy as np

pd.set_option("display.min_rows", 0) 
pd.set_option("display.max_rows", 30) 
pd.__version__

url = "https://drive.google.com/file/d/1afxwDXfl-7cQ_qLwyDitfcCx3u7WMvkU/view?usp=sharing" 
path = "https://drive.google.com/uc?export=download&id="+url.split("/")[-2]
pro = pd.read_csv(path)

      ===============================================

## Rename Columns , Set Index
- Rules: 
    - add 2-3 chars prefix to all columns
    - change column name to CamelCase
    - shorten it as max as possible
    - the Unique column shall have ALL_CAP name
 - ``` df.columns```   , ``` df.index``` 
 - ``` df=df.rename(columns={"A": "a", "B": "c"})``` 
 - ``` df.columns = ["a","b":"x"]``` 
     - take care, renaming the columns like that will convert the NAN to some value!!
 - ``` df=df.set_index("col")```  , ``` df=df.reset_index()``` 

pro.columns

pro=pro.rename(columns={"sku": "PRO_SKU"
                        , "name": "pro_Name"
                        , "desc": "pro_Desc"
                        , "price": "pro_Pr"
                        , "promo_price": "pro_PromPr"
                        , "in_stock": "pro_InStock"
                        , "type": "pro_Type"})

pro.columns

pro.index

    ==============================================================================================

## Explore the data
- ``` df.shape``` , ``` df.size``` , ``` df.ndim``` 
- ``` df.sample(5)``` , ``` df.info()``` 
- Numerical : ``` df.describe()``` , ``` df.col.nlargest()``` , ``` df.col.nsmallest()``` 
- Category : ``` df.nunique()``` , ``` df.unique() ``` 

pro.shape

pro.sample(10)
## price and promo_price require clean up TODO: fix

pro.info()  
## hint: there are no nulls 
## hint: price  is of type object, has to be float - TODO: fix required
## hint: promo_price is of type object, has to be float - TODO: fix required

pro.describe()
# in_stock seems to be bool TODO: convert to bool
# TODO: 75% are 0 !!! there is no stock!!

pro.nunique() # hint: id is unique per row 
# seems there is alot of duplicates

pro.pro_Pr.sort_values(ascending=False).unique()

pro.pro_PromPr.sort_values(ascending=False)

    ==============================================================================================

## Initial Clean
- Remember to create a copy of the df using ``` df.copy()``` 

pro_original = pro.copy()

      ===============================================

### Strip whitespaces
- ``` df.applymap(lambda x: x.strip() if isinstance(x, str) else x)```

pro = pro.applymap(lambda x: x.strip() if isinstance(x, str) else x)

      ===============================================
### Remove Duplicates Rows
- ``` df.duplicated().sum()``` 
- ``` df.loc[df.duplicated()==True]``` 
- ``` df=df.drop_duplicates() ``` 
- ``` df=df.drop_duplicates(subset=["col"])```  --> remove rows based on duplicated in specific column

pro.duplicated().sum() #hint : No duplicates

pro.loc[pro.duplicated()==True]

pro=pro.drop_duplicates()

pro.shape

pro.info()

      =======================

#### Remove duplicated rows related to Unique columns
- find possible duplicates ``` len(df.ol_ID.unique()) == df.shape[0] ``` 
- get the excat value for duplicate columns ``` df.loc[df.duplicated(subset="col")]``` 
- find all columns with same value ``` df[df.col=="val"]``` 

check if pro_SKU is unique

len(pro.PRO_SKU.unique()) == pro.shape[0]

pro.loc[pro.duplicated(subset="PRO_SKU")]

pro[pro.PRO_SKU=="APP1197"]
#since it will drop the second one, its okie

pro = pro.drop_duplicates(subset="PRO_SKU")

      ===============================================

### Drop duplicate/un-necessary Columns
- ``` df=df.drop(columns=["col1","col2"])``` 

pro.info()

pro_PromPr Column
    - Has no description
    - Not clear how it is related to other columns
    - Its name doesn't reflect any meaningfull data
    _ And finally the discount value would be concluded from the Ordelines Product_unit_Price (@ selling) vs the original Product price
Hence Its safe to drop it fully

pro = pro.drop(columns="pro_PromPr")

      ===============================================

### Focus : Bussines case --> filter un-necessary Data
- check if possible to shrink the table based on the bussines case we have
- ``` df=df.drop(columns=["col1","col2"])``` 

    ==============================================================================================

## Compare to original DataFrame
``` df.compare(df2)``` 

#pro.compare(pro_original)

    ==============================================================================================

## Export the cleaned DataFrame

``` df.to_pickle("file_name.pkl")``` 

pro.to_pickle("clean_Tables\Products_c.pkl")
