==> 3_Merge_analyze_Completed.ipynb <==
    ==============================================================================================

## Importing the DataFrames
- ``` df = pd.read_pickle("file_name.pkl")```  --> create 1 df from a pkl file

import pandas as pd
import numpy as np
pd.set_option("display.min_rows", 0) 
pd.set_option("display.max_rows", 30) 
pd.__version__

ord = pd.read_pickle("clean_Tables\Orders_c.pkl")
ord_s = pd.read_pickle("clean_Tables\Orders_s.pkl")
ol = pd.read_pickle("clean_Tables\OrderLines_c.pkl")
pro = pd.read_pickle("clean_Tables\Products_c.pkl")

ord.info()

ol.info()

pro.info()

    ==============================================================================================

## Merge DataFrames
- ``` df.shape``` , ``` df.size``` , ``` df.ndim``` 
- ``` df.sample(5)``` , ``` df.info()``` 
- Numerical : ``` df.describe()``` , ``` df.col.nlargest()``` , ``` df.col.nsmallest()``` 
- Category : ``` df.nunique()``` , ``` df.unique() ``` 

db = ord.copy()

db = db.merge(ol,how="inner",left_on="ORD_ID",right_on="ol_ORD_ID",validate="one_to_many")

db = db.merge(pro,how="inner",left_on="ol_Sku",right_on="PRO_SKU",validate="many_to_one")

pd.set_option('max_colwidth', 40)
pd.get_option('max_colwidth')

db.sample(5)

db_s = ord_s.copy()

db_s = db_s.merge(ol,how="inner",left_on="ORD_ID",right_on="ol_ORD_ID",validate="one_to_many")

db_s = db_s.merge(pro,how="inner",left_on="ol_Sku",right_on="PRO_SKU",validate="many_to_one")

pd.set_option('max_colwidth', 40)
pd.get_option('max_colwidth')

    ==============================================================================================

## Clean Merged Data
- Remember to create a copy of the df using ``` df.copy()``` 

db_original = db.copy()

      ===============================================

### Drop duplicated Columns, Reorder and Set Index
 - ``` df.columns```   , ``` df.index``` 
 - ``` df=df.rename(columns={"A": "a", "B": "c"})``` 
 - ``` df.columns = ["a","b":"x"]``` 
     - take care, renaming the columns like that will convert the NAN to some value!!
 - ``` df=df.set_index("col")```  , ``` df=df.reset_index()``` 

db = db.drop(columns=["ol_ORD_ID","ol_Sku","ord_State"])

db.info()

db.shape

db.index #hint: no need to change index

      ===============================================
      
### Clean NAN and empty cells
- ``` df.isna().sum()``` 
- check if it make sense to replace empty cells with NAN or with another value!
- ``` df = df.replace('^\s*$', np.nan)```  -->replace empty cells and cells with only whitspace with NAN
- check if other cells has the value you want to use to fill nan  ``` df.col.str.contains("pat", na=False).sum() ```
- ``` df=df.col.fillna(value,method="bfill"or"ffill",limit=value)``` 

- Extra: 
  - ``` (df.values == '').sum()```  --> check if any cell is empty
  - ``` df.col.str.isspace().sum()```  --> check if all cell is filled with whitespaces

db.info()

db.isna().sum()


db = db.replace('^\s*$', np.nan)
# TODO:HANA Question is it safe to do that?


db.isna().sum()

___ For pro_Desc   ___

db.pro_Desc.dtype

db.pro_Desc = db.pro_Desc.fillna("No Description")

db.isna().sum()

___ For pro_Type   ___

db.pro_Type.str.contains("^0+\.0+$", na=False).sum()
# No value is Zero , safely set NAN to 0

db.pro_Type = db.pro_Type.fillna("0")

db.isna().sum()

___ For pro_Pr   ___

db.pro_Pr.str.contains("^0+\.0+$", na=False).sum()
# No value is Zero , safely set NAN to 0

db.pro_Pr = db.pro_Pr.fillna("0")

db.isna().sum()

      ===============================================
### Fix DataTypes
- use ```df.nunique()``` to find if column is bool or Category
- before change, make sure all values will be converted correct
  - to convert to numeric : check no letters, 
      - if float(2dec) ```(~df.col.str.contains("^\d+.\d{2}$")).sum()```
      - if integer ```(~df.col.str.contains("^\d+$")).sum()```
  - to convert to category : check there are reasonable number of unique items
  - to convert to Datetime : ```(~df.col.str.contains("^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$")).sum()```
- ``` df.col.astype(type,errors="raise")``` 
  - type = "int","float","bool","category","object","datetime","timedelta"
- for mixed data
  - ``` pd.to_numeric(df.col, downcast=x,errors="raise") ``` 
  x = "integer" or "float"
  - ``` pd.to_datetime(df.col, downcast=None,errors="raise") ``` 
  - ``` pd.to_timedelta(df.col, downcast=None,errors="raise") ``` 

___ For ord_CreatDate ___

db.info()

(~db.ord_CreatDate.str.contains("^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$")).sum() 
# all rows has the datetime pattern, save to convert

db.ord_CreatDate=pd.to_datetime(db.ord_CreatDate,errors="raise")

db.info()

___ For ol_ProcessDate ___

(~db.ol_ProcessDate.str.contains("^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$")).sum() 
# all rows has the datetime pattern, save to convert

db.ol_ProcessDate=pd.to_datetime(db.ol_ProcessDate,errors="raise")

___ For pro_InStock ___

db.pro_InStock.unique()

db.pro_InStock = db.pro_InStock.astype("bool",errors="raise")

db.info()


___ For ol_ProdUntPr ___

#db_Pr = db[["ORD_ID","ord_TotlPaid","ol_ProdQnty","ol_ProdUntPr","pro_Pr","PRO_SKU","pro_Name"]].copy()


db.sample(5)

db.ol_ProdUntPr.dtype

(db.ol_ProdUntPr.str.count('\.') > 1).sum()

str_decide="Drop it"
x= round((((db.ol_ProdUntPr.str.count('\.') > 1).sum())/db.shape[0])*100,2)
if x > 5:
    str_decide = "Clean it"

print(f"the percentage of 2dec ol_ProdUntPr rows is {x}%, hence {str_decide}")

db[(db.ol_ProdUntPr.str.count('\.') > 1)].sample(5)

db=(
    db
    .assign(ol_ProdUntPr = lambda x:np.where(
      (x.ol_ProdUntPr.str.count('\.') == 2)
      ,x.ol_ProdUntPr.str.replace("\.","",1,regex=True)
      ,x.ol_ProdUntPr
      ))
)

db.info()

db.ol_ProdUntPr = db.ol_ProdUntPr.astype(float)

___ For pro_Pr  ___

db.sample(5)

str_decide="Drop it"
x= round((((db.pro_Pr.str.count('\.') > 1).sum())/db.shape[0])*100,2)
if x > 5:
    str_decide = "Clean it"
print(f"the percentage of 2dec pro_Pr rows is {round(x,2)}% , Hence {str_decide}")

db.shape

db = db.loc[db.pro_Pr.str.count('\.') < 2]

db.shape

db.pro_Pr = db.pro_Pr.astype(float)

db.info()

      ===============================================
### Analyze Outliers

remove un-necessary values(respectively related rows) of specific columns
- ``` Q1   = df.col.quantile(0.25) ``` 
- ``` Q3   = df.col.quantile(0.75) ``` 
- ``` IQR = Q3 - Q1 ``` 
- ``` num_inliers =((df.col >= (Q1 - 1.5*IQR)) & (df.col <= (Q3 + 1.5*IQR))).sum() ```
- ``` num_outliers = df.shape[0]-num_inliers ```
- ``` per_outliers = round((num_outliers/df.shape[0])*100,2) ```
- ``` print(f'''{num_outliers} out of {df.shape[0]} are outliers from col,ie {per_outliers}%''') ``` 
- ```if per_outliers < 3:```
    ```    print("Drop it")```
    ```else:```
    ```    print("Clean it") ```


db.describe()

__ ord_TotlPaid __

Q1   = db.ord_TotlPaid.quantile(0.25)
Q3   = db.ord_TotlPaid.quantile(0.75)
IQR = Q3 - Q1
num_inliers =((db.ord_TotlPaid >= (Q1 - 1.5*IQR)) & (db.ord_TotlPaid <= (Q3 + 1.5*IQR))).sum()
num_outliers = db.shape[0]-num_inliers
per_outliers = round((num_outliers/db.shape[0])*100,2)
print(f'''{num_outliers} out of {db.shape[0]} are outliers from ord_TotlPaid, ie {per_outliers}%''')
#df   = db.loc[]
if per_outliers < 3:
    print("Drop it")
else:
    print("Clean it")


db.describe()

    ==============================================================================================

## Re-Explore the data
draw some ``` df.col.hist()```  , ``` df.ser.boxplot()```  per column
take notes

db.columns

db.groupby(db.ord_CreatDate.dt.year).ord_CreatDate.hist()

db.ord_CreatDate.hist();

db.ord_TotlPaid.hist();

db.ol_ProdQnty.hist(bins=20);

db.ol_ProdUntPr.hist();

db.pro_Pr.hist();

db.pro_Type.hist();

db.columns

#db.groupby("ord_TotlPaid").boxplot()

    ==============================================================================================

## Export the cleaned DataFrame

``` df.to_pickle("file_name.pkl")``` 

db.to_pickle("clean_Tables\Merged_c.pkl")
